refactor: profile form data -> 공통 타입으로 수정
