+++ src/app/_types/profile.ts
@@ -6,3 +6,12 @@ export interface ConvertedProfileValues {
   dogImage?: boolean;
   dogCharacter?: boolean;
 }
+
+export interface ProfileFormData {
+  nickname: string;
+  email: string;
+  dogName: string;
+  dogBreed: string;
+  dogCharacter: string;
+  dogPersonality: string;
+}
